+++ pytest_inmanta/core.py
@@ -37,7 +37,7 @@ except DistributionNotFound:
 
 # Setting a project-wide pip index is only supported for iso7+
 SUPPORTS_PROJECT_PIP_INDEX: bool = (
-    CORE_VERSION is not None and CORE_VERSION >= version.Version("10.0.0.dev")
+    CORE_VERSION is not None and CORE_VERSION >= version.Version("9.0.0.dev")
 )
 
 SUPPORTS_MODULES_V2: bool = (