+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v 2.8.0 (?)
 Changes in this release:
 - Log a warning when the id_attribute of a resource is called id.
+- Ignore `__pycache__` dirs when copying the current module to the test project dir
 
 # v 2.7.0 (2023-02-23)
 Changes in this release: